Optimization

+++ Source/PCGExtendedToolkit/Public/Graph/PCGExEdgesProcessor.h
@@ -54,7 +54,7 @@ struct PCGEXTENDEDTOOLKIT_API FPCGExEdgesProcessorContext : public FPCGExPointsP
 
 	PCGExData::FPointIOTaggedDictionary* InputDictionary = nullptr;
 	PCGExData::FPointIOTaggedEntries* TaggedEdges = nullptr;
-	TMap<int32, int32> NodeIndicesMap;
+	TMap<int64, int32> NodeIndicesMap;
 	PCGEx::TFAttributeReader<int32>* EdgeNumReader = nullptr;
 
 	virtual bool AdvancePointsIO() override;
